Creating endpoints to the Category module and create with Doobie the dao layer for Category. And adding flyway for the migration of the database.

+++ src/main/scala/com/applaudo/akkalms/models/responses/FinanceResponse.scala
@@ -1,6 +1,6 @@
 package com.applaudo.akkalms.models.responses
 
-case class FinanceResponse(id: Int,
+case class FinanceResponse(id: Long,
                            year: Int,
                            month: String,
                            incomes: List[IncomeResponse])
